+++ src/service/sellerDashboard/SellerDashboardAPI.ts
@@ -1,5 +1,5 @@
 // src/services/SellerDashboardAPI.ts
-import {apiClient, retryIfUnauthorized} from '@/service/auth/AuthAPI';
+import {apiClient, retryIfUnauthorized} from '@/service/auth/AuthAPI.ts';
 
 // ===== 백엔드 응답 타입 정의 =====
